[add] style up

+++ src/client/components/Word.tsx
@@ -4,6 +4,8 @@ import { SyntheticEvent } from "react";
 import { connect } from "react-redux";
 import { updateWord, deleteWord } from "../actions";
 
+import styled, { css } from "react-emotion";
+
 interface IOwnProps {
   word: IWordModel;
 }
@@ -18,6 +20,26 @@ interface IDispatchProps {
   updateWord: (id: string, en: string, ko: string) => void;
   deleteWord: (id: string) => void;
 }
+
+const Button = styled("button")`
+  border: 0;
+  cursor: pointer;
+  color: white;
+  padding: .2rem .5rem;
+  background-color: ${props => props.theme.secondaryColor};
+  :hover {
+    background-color: #3498db;
+  }
+`;
+
+const Row = styled("tr")`
+  border-bottom: 1px solid black;
+  height: 5rem;
+`;
+
+const marginRight = css`
+  margin-right: 1rem;
+`
 
 type IProps = IOwnProps & IDispatchProps;
 
@@ -44,8 +66,10 @@ class Word extends React.Component<IProps, IState> {
   }
 
   private handleDelete() {
-    this.props.deleteWord(this.props.word._id)
-    this.handleEdit();
+    if (confirm("정말 삭제해도 되나요?")) {
+      this.props.deleteWord(this.props.word._id);
+      this.handleEdit();
+    }
   }
 
   private handleChangeEnglish(e: SyntheticEvent<HTMLInputElement>) {
@@ -65,7 +89,7 @@ class Word extends React.Component<IProps, IState> {
   public render() {
     if (this.state.editMode)
       return (
-        <tr>
+        <Row>
           <td>
             <input
               type="text"
@@ -80,24 +104,26 @@ class Word extends React.Component<IProps, IState> {
               onChange={this.handleChangeKorean.bind(this)}
             />
           </td>
-          <td>
-            <button onClick={this.handleApply.bind(this)}>Apply</button>
-            <button onClick={this.handleEdit.bind(this)}>Cancel</button>
-            <button onClick={this.handleDelete.bind(this)}>Delete</button>
+          <td style={{ whiteSpace: "nowrap" }}>
+            <Button className={marginRight} onClick={this.handleApply.bind(this)}>변경</Button>
+            <Button className={marginRight} onClick={this.handleEdit.bind(this)}>취소</Button>
+            <Button style={{backgroundColor: "red"}} onClick={this.handleDelete.bind(this)}>삭제</Button>
           </td>
-        </tr>
+        </Row>
       );
     else
       return (
-        <tr>
+        <Row>
           <td>{this.props.word.en}</td>
           <td>{this.props.word.ko}</td>
-          <td>
-            <button onClick={this.handleEdit.bind(this)}>edit</button>
+          <td style={{ whiteSpace: "nowrap" }}>
+            <Button onClick={this.handleEdit.bind(this)}>수정</Button>
           </td>
-        </tr>
+        </Row>
       );
   }
 }
 
-export default connect<{}, IDispatchProps>(null, { updateWord, deleteWord })(Word);
+export default connect<{}, IDispatchProps>(null, { updateWord, deleteWord })(
+  Word,
+);
